requirements.txt
pandas
numpy
tqdm
dvc[s3]
ipykernel
pillow
keras
tensorflow